requirements.txt
pydantic>=2
rich
jinja2
pyyaml
python-dotenv
requests
reportlab
